Move app to root/ folder
