+++ web/app/root/root.component.ts
@@ -1,6 +1,6 @@
 import {Component, HostBinding} from '@angular/core';
 
-@Component({selector: 'fci-root', templateUrl: './app.component.html'})
-export class AppComponent {
+@Component({selector: 'fci-root', templateUrl: './root.component.html'})
+export class RootComponent {
   @HostBinding('class') classes = ['fci-full-height-container'];
 }
