+++ web/app/root/routing.module.ts
@@ -1,14 +1,14 @@
 import {NgModule} from '@angular/core';
 import {RouterModule, Routes} from '@angular/router';
 
-import {BuildComponent} from './build/build.component';
-import {BuildModule} from './build/build.module';
-import {DashboardComponent} from './dashboard/dashboard.component';
-import {DashboardModule} from './dashboard/dashboard.module';
-import {LoginComponent} from './login/login.component';
-import {LoginModule} from './login/login.module';
-import {ProjectComponent} from './project/project.component';
-import {ProjectModule} from './project/project.module';
+import {BuildComponent} from '../build/build.component';
+import {BuildModule} from '../build/build.module';
+import {DashboardComponent} from '../dashboard/dashboard.component';
+import {DashboardModule} from '../dashboard/dashboard.module';
+import {LoginComponent} from '../login/login.component';
+import {LoginModule} from '../login/login.module';
+import {ProjectComponent} from '../project/project.component';
+import {ProjectModule} from '../project/project.module';
 
 const routes: Routes = [
   {path: '', redirectTo: '/dashboard', pathMatch: 'full'},
@@ -26,5 +26,5 @@ const routes: Routes = [
   exports: [RouterModule]
 })
 
-export class AppRoutingModule {
+export class RoutingModule {
 }
